shoutouts

+++ app/db/db_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, TIMESTAMP, BOOLEAN, ForeignKey, FLOAT
+from sqlalchemy import Column, BigInteger, String, TIMESTAMP, BOOLEAN, ForeignKey, FLOAT
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from app.db.session import Base
@@ -7,22 +7,22 @@ from datetime import datetime
 
 class User(Base):
     __tablename__ = "users"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, unique=True, nullable=False)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, unique=True, nullable=False)
     uuid = Column("uuid", UUID(as_uuid=True), primary_key=True, unique=True, nullable=False)
     googleId = Column("google_id", String, unique=True)
     vkId = Column("vk_id", String, unique=True)
     appleId = Column("apple_id", String, unique=True)
     email = Column("email", String, unique=True)
     emailVerified = Column("email_verified", BOOLEAN, nullable=False)
-    phone = Column("phone", Integer, unique=True)
+    phone = Column("phone", BigInteger, unique=True)
     phoneVerified = Column("phone_verified", BOOLEAN, nullable=False)
     password = Column("password", String)
     name = Column("name", String)
     surname = Column("surname", String)
     username = Column("username", String, unique=True)
-    rating = Column("rating", Integer)
-    photoId = Column("photo_id", Integer, ForeignKey("photos.id"))
-    roleId = Column("roleId", Integer, ForeignKey("roles.id"), nullable=False)
+    rating = Column("rating", BigInteger)
+    photoId = Column("photo_id", BigInteger, ForeignKey("photos.id"))
+    roleId = Column("roleId", BigInteger, ForeignKey("roles.id"), nullable=False)
     createdAt = Column("created_at", TIMESTAMP, nullable=False)
     updatedAt = Column("updated_at", TIMESTAMP)
     lastLoginAt = Column("last_login_at", TIMESTAMP)
@@ -40,7 +40,7 @@ class User(Base):
 
 class Photo(Base):
     __tablename__ = "photos"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
     url = Column("url", String)
 
     owner = relationship("User", back_populates="photo")
@@ -48,7 +48,7 @@ class Photo(Base):
 
 class Role(Base):
     __tablename__ = "roles"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
     title = Column("title", String)
 
     owner = relationship("User", back_populates="role")
@@ -56,47 +56,47 @@ class Role(Base):
 
 class Post(Base):
     __tablename__ = "posts"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
-    random = Column("random", Integer)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
+    random = Column("random", BigInteger)
     title = Column("title", String)
     description = Column("description", String)
-    priceAmount = Column("price_amount", Integer)
+    priceAmount = Column("price_amount", BigInteger)
     priceCurrency = Column("price_currency", String)
     link = Column("link", String)
     date = Column("date", TIMESTAMP)
     source = Column("source", String)
     parseDate = Column("parse_date", TIMESTAMP)
-    priority = Column("priority", Integer)
+    priority = Column("priority", BigInteger)
 
 
 class Jobs(Base):
     __tablename__ = "jobs"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
-    posts_parsed = Column("posts_parsed", Integer)
-    posts_uploaded = Column("posts_uploaded", Integer)
-    posts_update = Column("posts_update", Integer)
-    posts_delete = Column("posts_delete", Integer)
-    errors = Column("errors", Integer)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
+    posts_parsed = Column("posts_parsed", BigInteger)
+    posts_uploaded = Column("posts_uploaded", BigInteger)
+    posts_update = Column("posts_update", BigInteger)
+    posts_delete = Column("posts_delete", BigInteger)
+    errors = Column("errors", BigInteger)
     lead_time = Column("lead_time", FLOAT)
     date = Column("date", TIMESTAMP)
 
 
 class Errors(Base):
     __tablename__ = "errors"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
     error = Column("error", String)
     description = Column("description", String)
     source = Column("source", String)
-    job_id = Column("job_id", Integer)
+    job_id = Column("job_id", BigInteger)
 
 
 class Vacancy(Base):
     __tablename__ = "vacancies"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
-    userId = Column("user_id", Integer, ForeignKey("users.id"), nullable=False)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
+    userId = Column("user_id", BigInteger, ForeignKey("users.id"), nullable=False)
     title = Column("title", String(300), nullable=False)
     description = Column("description", String(6000))
-    budget = Column("budget", Integer, default=None)
+    budget = Column("budget", BigInteger, default=None)
     name = Column("name", String)
     email = Column("email", String)
     phone = Column("phone", String)
@@ -107,25 +107,25 @@ class Vacancy(Base):
 
 class File(Base):
     __tablename__ = "files"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
     filename = Column("filename", String)
     patch = Column("patch", String)
-    vacancyId = Column("vacancy_id", Integer, ForeignKey("vacancies.id"))
+    vacancyId = Column("vacancy_id", BigInteger, ForeignKey("vacancies.id"))
 
 
 class Favorites(Base):
     __tablename__ = "favorites"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
-    userId = Column("user_id", Integer, ForeignKey("users.id"))
-    objId = Column("obj_id", Integer, ForeignKey("posts.id"))
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
+    userId = Column("user_id", BigInteger, ForeignKey("users.id"))
+    objId = Column("obj_id", BigInteger, ForeignKey("posts.id"))
 
 
 class Shoutout(Base):
     __tablename__ = "shoutout"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
-    userId = Column("user_id", Integer, ForeignKey("users.id"))
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
+    userId = Column("user_id", BigInteger, ForeignKey("users.id"))
     text = Column("text", String, nullable=False)
-    rating = Column("rating", Integer, nullable=False)
+    rating = Column("rating", BigInteger, nullable=False)
     createdAt = Column("created_at", TIMESTAMP)
 
     user = relationship("User", back_populates="shoutouts")
@@ -133,9 +133,9 @@ class Shoutout(Base):
 
 class HistorySearch(Base):
     __tablename__ = "historySearch"
-    id = Column("id", Integer, primary_key=True, index=True, autoincrement=True, nullable=False)
-    userId = Column("user_id", Integer, ForeignKey("users.id"))
+    id = Column("id", BigInteger, primary_key=True, index=True, autoincrement=True, nullable=False)
+    userId = Column("user_id", BigInteger, ForeignKey("users.id"))
     searchQuery = Column("search_query", String)
     flagPrice = Column("flag_price", BOOLEAN, default=False)
-    price = Column("price", Integer, default=0)
+    price = Column("price", BigInteger, default=0)
     date = Column("date", TIMESTAMP, default=datetime.now())
